\name{EPAM110}
\docType{data}
\alias{EBLOSUM40}
\alias{EPAM300}
\alias{EPAM110}
\alias{EBLOSUM30}
\alias{EBLOSUM35} 
\alias{EBLOSUM45}
\alias{EBLOSUM50}
\alias{EBLOSUM55} 
\alias{EBLOSUM60}
\alias{EBLOSUM62}
\alias{EBLOSUM65}
\alias{EBLOSUM70}
\alias{EBLOSUM75}
\alias{EBLOSUM80}
\alias{EBLOSUM85}
\alias{EBLOSUM90}
\alias{EPAM10}
\alias{EPAM100}
\alias{EPAM120}
\alias{EPAM130}
\alias{EPAM140}
\alias{EPAM150}
\alias{EPAM160}
\alias{EPAM170}
\alias{EPAM180}
\alias{EPAM190}
\alias{EPAM20}
\alias{EPAM200}
\alias{EPAM210}
\alias{EPAM220}
\alias{EPAM230}
\alias{EPAM240}
\alias{EPAM250}
\alias{EPAM260}
\alias{EPAM270}
\alias{EPAM280}
\alias{EPAM290}
\alias{EPAM30}
\alias{EPAM310}
\alias{EPAM320}
\alias{EPAM330}
\alias{EPAM340}
\alias{EPAM350}
\alias{EPAM360}
\alias{EPAM370}
\alias{EPAM380}
\alias{EPAM390}
\alias{EPAM40}
\alias{EPAM400}
\alias{EPAM410}
\alias{EPAM420}
\alias{EPAM430}
\alias{EPAM440}
\alias{EPAM450}
\alias{EPAM460}
\alias{EPAM470}
\alias{EPAM480}
\alias{EPAM490}
\alias{EPAM50}
\alias{EPAM500}
\alias{EPAM60}
\alias{EPAM70}
\alias{EPAM80}
\alias{EPAM90}
\title{Amino Acid Substitution matrices}
\description{
Amino Acid Substitution matrices
}
\usage{data(EPAM110)}
\format{}
\details{
Amino Acid Substitution Matrices.
}
\source{
\url{http://www.hgmp.mrc.ac.uk/Software/EMBOSS/}
}
\references{http://r4protoemics.sourceforge.net}
}
\examples{
data(EPAM110)
}
\keyword{datasets}
