\name{sequlist}
\alias{sequlist}
\docType{data}
\title{AASequenceList with 66 entries.}
\description{
  66 arbitrary entries from the \emph{ arabidopsis thaliana} sequence database (arabi\_all\_proteins\_v110402.fasta). 
}
\usage{data(sequlist)}
\format{
}
\details{
}
\source{
  \url{}
}
\references{\url{http://r4proteomics.sourceforge.net}}
\examples{
data(sequlist)
## maybe str(sequlist) ; plot(sequlist) ...
length(sequlist)
}
\keyword{datasets}
